packages/core: rename api definitions to CamelCaseApi
