+++ packages/core/src/api/apis/definitions/ErrorApi.ts
