+++ packages/core/src/api/apis/definitions/OAuthRequestApi.ts
