+++ packages/core/src/api/apis/definitions/index.ts
@@ -20,7 +20,7 @@
 //
 // If you think some API definition is missing, please open an Issue or send a PR!
 
-export * from './alert';
-export * from './error';
-export * from './featureFlags';
-export * from './oauthrequest';
+export * from './AlertApi';
+export * from './ErrorApi';
+export * from './FeatureFlagsApi';
+export * from './OAuthRequestApi';
